filter input

+++ app/components/delete-post.tsx
@@ -1,6 +1,7 @@
 "use client";
 import { API_URL } from "@/config/site";
 import axios from "axios";
+import { FaTrash } from "react-icons/fa";
 
 interface DeletePostProps {
   postId: string;
@@ -9,10 +10,11 @@ interface DeletePostProps {
 
 export default function DeletePost({ postId, onDelete }: DeletePostProps) {
   const handleDelete = async () => {
+    const isConfirmed = confirm("¿Deseas eliminar esta publicación?");
+    if (!isConfirmed) return;
     try {
       const response = await axios.delete(API_URL + `/posts/${postId}`);
       if (response.status === 200) {
-        alert("Post eliminado");
         onDelete();
       }
     } catch (error) {
@@ -21,8 +23,11 @@ export default function DeletePost({ postId, onDelete }: DeletePostProps) {
   };
 
   return (
-    <button onClick={handleDelete} className="bg-red-500 text-white px-4 py-2 rounded">
-      Eliminar
+    <button
+      onClick={handleDelete}
+      className="bg-red-500 text-white font-semibold px-4 py-2 rounded flex items-center gap-2"
+    >
+      Eliminar <FaTrash />
     </button>
   );
 }
